Move interface to shell/platform/android/jni/
