+++ shell/platform/android/android_shell_holder.h
@@ -14,8 +14,8 @@
 #include "flutter/shell/common/run_configuration.h"
 #include "flutter/shell/common/shell.h"
 #include "flutter/shell/common/thread_host.h"
+#include "flutter/shell/platform/android/jni/platform_view_android_jni.h"
 #include "flutter/shell/platform/android/platform_view_android.h"
-#include "flutter/shell/platform/android/platform_view_android_jni.h"
 
 namespace flutter {
