+++ shell/platform/android/android_surface_software.h
@@ -11,7 +11,7 @@
 #include "flutter/shell/gpu/gpu_surface_software.h"
 #include "flutter/shell/platform/android/android_surface.h"
 #include "flutter/shell/platform/android/external_view_embedder/external_view_embedder.h"
-#include "flutter/shell/platform/android/platform_view_android_jni.h"
+#include "flutter/shell/platform/android/jni/platform_view_android_jni.h"
 
 namespace flutter {
