+++ shell/platform/android/android_surface_vulkan.h
@@ -12,7 +12,7 @@
 #include "flutter/shell/platform/android/android_native_window.h"
 #include "flutter/shell/platform/android/android_surface.h"
 #include "flutter/shell/platform/android/external_view_embedder/external_view_embedder.h"
-#include "flutter/shell/platform/android/platform_view_android_jni.h"
+#include "flutter/shell/platform/android/jni/platform_view_android_jni.h"
 #include "flutter/vulkan/vulkan_window.h"
 
 namespace flutter {
