+++ shell/platform/android/external_view_embedder/BUILD.gn
@@ -20,7 +20,7 @@ source_set("external_view_embedder") {
 
   deps = [
     "//flutter/common",
-    "//flutter/shell/platform/android:android_jni",
+    "//flutter/shell/platform/android/jni",
     "//flutter/flow",
     "//flutter/fml",
     "//third_party/skia",
