+++ shell/platform/android/jni/BUILD.gn
@@ -1,0 +1,20 @@
+# Copyright 2013 The Flutter Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+import("//flutter/common/config.gni")
+
+source_set("jni") {
+  sources = [
+    "platform_view_android_jni.cc",
+    "platform_view_android_jni.h",
+  ]
+
+  public_configs = [ "//flutter:config" ]
+
+  deps = [
+    "//flutter/fml",
+    "//flutter/lib/ui",
+    "//third_party/skia",
+  ]
+}
