+++ shell/platform/android/platform_view_android_jni_impl.h
@@ -6,7 +6,7 @@
 #define FLUTTER_SHELL_PLATFORM_ANDROID_PLATFORM_VIEW_ANDROID_JNI_IMPL_H_
 
 #include "flutter/fml/platform/android/jni_weak_ref.h"
-#include "flutter/shell/platform/android/platform_view_android_jni.h"
+#include "flutter/shell/platform/android/jni/platform_view_android_jni.h"
 
 namespace flutter {
